feat: added RLP tests
